feat: Add reload_marketing4than.service event listener
Add a new event listener to reload the marketing4than.service when the "reload_marketing4than.service" event is received from the server Socket.IO. This event is triggered when there are updates to the service. The listener checks the platform and uses the appropriate command to reload the service on Linux systems.